+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Failed-record error messages should use detailed "message", when present, instead of label. Fixes UIHAADM-125.
 * Verify that stats-message parsing works for new format, add test suite. Fixes UIHAADM-124.
 * When sending end-range dates as part of a query, extend them to the datetimes at the end of the day. Fixes UIHAADM-126.
+* Close button on last-logs page returns to harvestables list or detail page depending on where it came from. Fixes UIHAADM-129.
 
 ## [2.1.0](https://github.com/folio-org/ui-harvester-admin/tree/v2.1.0) (2024-02-28)